fix : column type affectation to work with postgres

+++ Staging/dbtStaging/models/certdc/ods_insee.sql
@@ -8,7 +8,7 @@ WITH sa_insee_total AS (
         *
         , 1 AS Total
     FROM {{ ref('staging__sa_insee_histo') }}
-    WHERE adec = {{ get_previous_year(3) }} OR adec = {{ get_previous_year(2) }}
+    WHERE adec::INT = {{ get_previous_year(3) }} OR adec::INT = {{ get_previous_year(2) }}
     UNION
     SELECT 
         adec,
